remove: redundant code
